[Refactor] Video grid (#4466)
* Pull GridView responsibilities up to VideoGridViewController

* Clean up VideoGridViewController

* Fix tests

* Clean up

* Fix grid again

* Restore fill mode switching

* Clean up

* Update snapshots

* Disable collection view animations

* Don't assume a single section

* Rename case

* Update Wire-iOS/Sources/UserInterface/Calling/VideoGridView/GridView.swift
